Update on Jan, 18th.

+++ src/com/manu/dynasty/template/Action.java
@@ -3,11 +3,11 @@ package com.manu.dynasty.template;
 public class Action {
 	public int Id;
 	public int TypeKey;
-	public int Param1;
-	public int Param2;
-	public int Param3;
-	public int Param4;
-	public int Param5;
+	public float Param1;
+	public float Param2;
+	public float Param3;
+	public float Param4;
+	public float Param5;
 	public int TSR;
 	public int TTR;
 	public int CeOnHit;
